CORE-7209 Create and use EmptyStringValueChangeHandler
